ZBUG-3235: allow a single character sub domain for the forwarding address
[Bug description]
A forwarding address whose subdomain consists one character is rejected
to input in the field of Admin Console -> Manage -> Accounts -> (pick
one account) -> Forwarding -> User-spefieid forwarding addresses.
"Invalid email address list" in red is displayed.

[Root cause]
The regex `AjxUtil.EMAIL_FULL_RE` for checking the
ZimbraPrefMailForwardingAddress attribute had been set not to allow a single
character subdomain.

According to the RFC 1034, the length of the subdomain is defined as
"... is zero to 63 octets in length." (practically it's 1 to 63
octets).

[Fix]
Changed the regex `AjxUtil.EMAIL_FULL_RE` to allow one character
subdomain.